Clean directory bin lib divide
